Adding pbs_sched_bare, minimal sched for testing
