+++ src/scheduler/Makefile.am
@@ -97,6 +97,7 @@ libpbs_sched_a_SOURCES = \
 	parse.h \
 	pbs_bitmap.cpp \
 	pbs_bitmap.h \
+	pbs_sched_utils.cpp \
 	prev_job_info.cpp \
 	prev_job_info.h \
 	prime.cpp \
@@ -124,10 +125,16 @@ libpbs_sched_a_SOURCES = \
 	site_data.h
 
 sbin_PROGRAMS = pbs_sched pbsfs
+unsupporteddir = ${exec_prefix}/unsupported
+unsupported_PROGRAMS = pbs_sched_bare
 
 pbs_sched_CPPFLAGS = ${common_cflags}
 pbs_sched_LDADD = ${common_libs} @libundolr_lib@
 pbs_sched_SOURCES = pbs_sched.cpp
+
+pbs_sched_bare_CPPFLAGS = ${common_cflags}
+pbs_sched_bare_LDADD = ${common_libs} @libundolr_lib@
+pbs_sched_bare_SOURCES = pbs_sched_bare.cpp
 
 pbsfs_CPPFLAGS = ${common_cflags}
 pbsfs_LDADD = ${common_libs}
